Removed (pointless?) casts
git-svn-id: https://svn.blender.org/svnroot/verse/branches/release-r3-branch@313 49c17534-1d02-dc11-b283-0030488c597c

+++ vs_node_bitmap.c
@@ -383,31 +383,31 @@ static void callback_send_b_layer_subscribe(void *user, VNodeID node_id, VLayerI
 		for(i = 0; i < tile[0]; i++)
 			for(j = 0; j < tile[1]; j++)
 				for(k = 0; k < tile[2]; k++)
-					verse_send_b_layer_set_tile(node_id, layer_id, (uint16)i, (uint16)j, (uint16)k, VN_B_LAYER_UINT1, &((VNBTile *)node->layers[layer_id].layer)[(tile[0] * tile[1] * k + j * tile[0] + i) * VN_B_TILE_SIZE * VN_B_TILE_SIZE / 8]);
+					verse_send_b_layer_set_tile(node_id, layer_id, i, j, k, VN_B_LAYER_UINT1, &((VNBTile *)node->layers[layer_id].layer)[(tile[0] * tile[1] * k + j * tile[0] + i) * VN_B_TILE_SIZE * VN_B_TILE_SIZE / 8]);
 	break;
 	case VN_B_LAYER_UINT8 :
 		for(i = 0; i < tile[0]; i++)
 			for(j = 0; j < tile[1]; j++)
 				for(k = 0; k < tile[2]; k++)
-					verse_send_b_layer_set_tile(node_id, layer_id, (uint16)i, (uint16)j, (uint16)k, VN_B_LAYER_UINT8, &((VNBTile *)node->layers[layer_id].layer)[(tile[0] * tile[1] * k + j * tile[0] + i) * VN_B_TILE_SIZE * VN_B_TILE_SIZE]);
+					verse_send_b_layer_set_tile(node_id, layer_id, i, j, k, VN_B_LAYER_UINT8, &((VNBTile *)node->layers[layer_id].layer)[(tile[0] * tile[1] * k + j * tile[0] + i) * VN_B_TILE_SIZE * VN_B_TILE_SIZE]);
 	break;
 	case VN_B_LAYER_UINT16 :
 		for(i = 0; i < tile[0]; i++)
 			for(j = 0; j < tile[1]; j++)
 				for(k = 0; k < tile[2]; k++)
-					verse_send_b_layer_set_tile(node_id, layer_id, (uint16)i, (uint16)j, (uint16)k, VN_B_LAYER_UINT16, &((VNBTile *)node->layers[layer_id].layer)[(tile[0] * tile[1] * k + j * tile[0] + i) * VN_B_TILE_SIZE * VN_B_TILE_SIZE]);
+					verse_send_b_layer_set_tile(node_id, layer_id, i, j, k, VN_B_LAYER_UINT16, &((VNBTile *)node->layers[layer_id].layer)[(tile[0] * tile[1] * k + j * tile[0] + i) * VN_B_TILE_SIZE * VN_B_TILE_SIZE]);
 	break;
 	case VN_B_LAYER_REAL32 :
 		for(i = 0; i < tile[0]; i++)
 			for(j = 0; j < tile[1]; j++)
 				for(k = 0; k < tile[2]; k++)
-					verse_send_b_layer_set_tile(node_id, layer_id, (uint16)i, (uint16)j, (uint16)k, VN_B_LAYER_REAL32, &((VNBTile *)node->layers[layer_id].layer)[(tile[0] * tile[1] * k + j * tile[0] + i) * VN_B_TILE_SIZE * VN_B_TILE_SIZE]);
+					verse_send_b_layer_set_tile(node_id, layer_id, i, j, k, VN_B_LAYER_REAL32, &((VNBTile *)node->layers[layer_id].layer)[(tile[0] * tile[1] * k + j * tile[0] + i) * VN_B_TILE_SIZE * VN_B_TILE_SIZE]);
 	break;
 	case VN_B_LAYER_REAL64 :
 		for(i = 0; i < tile[0]; i++)
 			for(j = 0; j < tile[1]; j++)
 				for(k = 0; k < tile[2]; k++)
-					verse_send_b_layer_set_tile(node_id, layer_id, (uint16)i, (uint16)j, (uint16)k, VN_B_LAYER_REAL64, &((VNBTile *)node->layers[layer_id].layer)[(tile[0] * tile[1] * k + j * tile[0] + i) * VN_B_TILE_SIZE * VN_B_TILE_SIZE]);
+					verse_send_b_layer_set_tile(node_id, layer_id, i, j, k, VN_B_LAYER_REAL64, &((VNBTile *)node->layers[layer_id].layer)[(tile[0] * tile[1] * k + j * tile[0] + i) * VN_B_TILE_SIZE * VN_B_TILE_SIZE]);
 	break;
 	}
 }
